+++ docs/dev/CHANGELOG.md
@@ -13,6 +13,8 @@ Released on: ???
 - Added Octopus Tendril and Troubled Bubble rare drop alert & party chat message. Make sure to enable if you need it! (or use ALL)
 - Added toggle to use gradient color for sea creature name in "Compact sea creature catch messages" feature.
 - Gradient color for mod name prefix in chat c:
+- Added Flux and its variations to Deployable overlay / Deployable expiration alert.
+- Added setting to control amount of seconds for alert before deployable expiration.
 
 ## Bugfixes